tech-debt: split lineup contract schema and extract shared PHASES constant (ROK-975)
Split lineup.schema.ts (337 lines) into two files to comply with the
300-line limit. Match/scheduling enums and DTOs (ROK-964) are extracted
to lineup-match.schema.ts, with re-exports from lineup.schema.ts for
backward compatibility.

Also extract duplicate PHASES/PHASE_LABELS constants from
LineupDetailHeader and LineupProgressBar into a shared lineup-phases.ts.
